+++ gateway/src/backend/events/mod.rs
@@ -1,3 +1,5 @@
+mod from;
+
 use std::time::SystemTime;
 
 use serde_derive::{Deserialize, Serialize};