+++ gateway/src/redis_io/mod.rs
@@ -24,15 +24,27 @@ pub fn create_pool() -> Arc<RedisPool> {
 
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct AllEntryIds {
-    pub bot_attached_eid: XReadEntryId,
-    pub move_made_eid: XReadEntryId,
+    pub bot_attached_xid: XReadEntryId,
+    pub move_made_xid: XReadEntryId,
+    pub hist_prov_xid: XReadEntryId,
+    pub sync_reply_xid: XReadEntryId,
+    pub wait_opponent_xid: XReadEntryId,
+    pub game_ready_xid: XReadEntryId,
+    pub priv_game_reject_xid: XReadEntryId,
+    pub colors_chosen_xid: XReadEntryId,
 }
 
 impl Default for AllEntryIds {
     fn default() -> Self {
         AllEntryIds {
-            bot_attached_eid: XReadEntryId::default(),
-            move_made_eid: XReadEntryId::default(),
+            bot_attached_xid: XReadEntryId::default(),
+            move_made_xid: XReadEntryId::default(),
+            hist_prov_xid: XReadEntryId::default(),
+            sync_reply_xid: XReadEntryId::default(),
+            wait_opponent_xid: XReadEntryId::default(),
+            game_ready_xid: XReadEntryId::default(),
+            priv_game_reject_xid: XReadEntryId::default(),
+            colors_chosen_xid: XReadEntryId::default(),
         }
     }
 }