+++ gateway/src/redis_io/stream.rs
@@ -1,10 +1,27 @@
 use super::entry_id_repo::*;
 use super::xread::XReader;
-use crate::backend_events::BackendEvents;
-use crate::model::{Coord, MoveMadeEvent, Player};
-
+use crate::backend::events as be;
+use crate::backend::events::BackendEvents;
+use crate::model::{ColorsChosenEvent, HistoryProvidedEvent, MoveMadeEvent};
+use color_model as color;
 use crossbeam_channel::Sender;
-use log::error;
+use lobby_model as lobby;
+use log::{error, info};
+use move_model as moves;
+use redis_streams::XReadEntryId;
+use sync_model as sync;
+
+#[derive(Clone, Debug)]
+pub enum StreamData {
+    BotAttached(micro_model_bot::gateway::BotAttached),
+    MoveMade(moves::MoveMade),
+    HistoryProvided(sync::api::HistoryProvided),
+    SyncReply(sync::api::SyncReply),
+    WaitForOpponent(lobby::api::WaitForOpponent),
+    GameReady(lobby::api::GameReady),
+    PrivGameRejected(lobby::api::PrivateGameRejected),
+    ColorsChosen(color::api::ColorsChosen),
+}
 
 pub fn process(events_in: Sender<BackendEvents>, opts: StreamOpts) {
     loop {
@@ -13,59 +30,29 @@ pub fn process(events_in: Sender<BackendEvents>, opts: StreamOpts) {
             Ok(entry_ids) => match opts.xreader.xread_sorted(entry_ids) {
                 Err(e) => error!("cannot xread {:?}", e),
                 Ok(xrr) => {
-                    for event in xrr {
-                        match event {
-                            (entry_id, StreamData::BotAttached(b)) => {
-                                if let Err(e) = events_in.send(BackendEvents::BotAttached(b)) {
-                                    error!("send err bot attached {:?}", e)
-                                } else if let Err(e) = opts
-                                    .entry_id_repo
-                                    .update(EntryIdType::BotAttached, entry_id)
-                                {
-                                    error!("err tracking EID bot attached {:?}", e)
-                                }
-                            }
-                            (
-                                entry_id,
-                                StreamData::MoveMade(micro_model_moves::MoveMade {
-                                    game_id,
-                                    coord,
-                                    reply_to,
-                                    player,
-                                    captured,
-                                    event_id,
-                                }),
-                            ) => {
-                                if let Err(e) =
-                                    events_in.send(BackendEvents::MoveMade(MoveMadeEvent {
-                                        game_id: game_id.0,
-                                        coord: coord.map(|c| Coord { x: c.x, y: c.y }),
-                                        reply_to: reply_to.0,
-                                        player: match player {
-                                            micro_model_moves::Player::BLACK => Player::BLACK,
-                                            _ => Player::WHITE,
-                                        },
-                                        captured: captured
-                                            .iter()
-                                            .map(|c| Coord { x: c.x, y: c.y })
-                                            .collect(),
-                                        event_id: event_id.0,
-                                    }))
-                                {
-                                    error!("send err move made {:?}", e)
-                                }
-
-                                if let Err(e) =
-                                    opts.entry_id_repo.update(EntryIdType::MoveMade, entry_id)
-                                {
-                                    error!("err tracking EID move made {:?}", e)
-                                }
-                            }
-                        }
+                    for (xid, data) in xrr {
+                        info!("📥 Stream: {:?}", &data);
+                        let dc = data.clone();
+                        process_event(xid, data, &events_in, &opts);
+                        info!("🏞 OK {:?}", dc)
                     }
                 }
             },
         }
+    }
+}
+
+fn process_event(
+    xid: XReadEntryId,
+    data: StreamData,
+    events_in: &Sender<BackendEvents>,
+    opts: &StreamOpts,
+) {
+    let xid_type = EntryIdType::from(&data);
+    if let Err(e) = events_in.send(BackendEvents::from(data)) {
+        error!("send backend event {:?}", e)
+    } else if let Err(e) = opts.entry_id_repo.update(xid_type, xid) {
+        error!("err tracking XID {:?}", e)
     }
 }
 
@@ -74,8 +61,44 @@ pub struct StreamOpts {
     pub xreader: Box<dyn XReader>,
 }
 
-#[derive(Clone, Debug)]
-pub enum StreamData {
-    BotAttached(micro_model_bot::gateway::BotAttached),
-    MoveMade(micro_model_moves::MoveMade),
+impl From<&StreamData> for EntryIdType {
+    fn from(data: &StreamData) -> Self {
+        match data {
+            StreamData::BotAttached(_) => EntryIdType::BotAttached,
+            StreamData::MoveMade(_) => EntryIdType::MoveMade,
+            StreamData::HistoryProvided(_) => EntryIdType::HistProv,
+            StreamData::SyncReply(_) => EntryIdType::SyncReply,
+            StreamData::WaitForOpponent(_) => EntryIdType::WaitOpponent,
+            StreamData::GameReady(_) => EntryIdType::GameReady,
+            StreamData::PrivGameRejected(_) => EntryIdType::PrivGameReject,
+            StreamData::ColorsChosen(_) => EntryIdType::ColorsChosen,
+        }
+    }
+}
+
+impl From<StreamData> for BackendEvents {
+    fn from(stream_data: StreamData) -> Self {
+        match stream_data {
+            StreamData::MoveMade(m) => BackendEvents::MoveMade(MoveMadeEvent::from(m)),
+            StreamData::BotAttached(b) => BackendEvents::BotAttached(b),
+            StreamData::HistoryProvided(h) => {
+                BackendEvents::HistoryProvided(HistoryProvidedEvent::from(h))
+            }
+            StreamData::SyncReply(s) => {
+                BackendEvents::SyncReply(be::SyncReplyBackendEvent::from(s))
+            }
+            StreamData::WaitForOpponent(w) => {
+                be::BackendEvents::WaitForOpponent(be::WaitForOpponentBackendEvent::from(w))
+            }
+            StreamData::GameReady(g) => {
+                be::BackendEvents::GameReady(be::GameReadyBackendEvent::from(g))
+            }
+            StreamData::PrivGameRejected(p) => {
+                be::BackendEvents::PrivateGameRejected(be::PrivateGameRejectedBackendEvent::from(p))
+            }
+            StreamData::ColorsChosen(c) => {
+                be::BackendEvents::ColorsChosen(ColorsChosenEvent::from(c))
+            }
+        }
+    }
 }