+++ gateway/src/redis_io/xadd.rs
@@ -1,31 +1,54 @@
-use crate::backend_commands::BackendCommands;
-use crate::model::{Coord, MakeMoveCommand};
+use crate::backend::commands::BackendCommands as BC;
+use crate::backend::commands::{
+    ChooseColorPrefBackendCommand, CreateGameBackendCommand, FindPublicGameBackendCommand,
+    JoinPrivateGameBackendCommand, ReqSyncBackendCommand, SessionDisconnected,
+};
+use crate::model::{Coord, MakeMoveCommand, ProvideHistoryCommand};
 use crate::redis_io::RedisPool;
-use crate::topics::{ATTACH_BOT_TOPIC, MAKE_MOVE_TOPIC};
+use crate::topics;
 use micro_model_bot::gateway::AttachBot;
 
+use crate::backend::commands::IntoShared;
 use crossbeam_channel::{select, Receiver};
-use log::error;
+use log::{error, info};
 use r2d2_redis::redis;
 use std::sync::Arc;
 
 pub trait XAddCommands {
     fn xadd_attach_bot(&self, attach_bot: AttachBot);
     fn xadd_make_move(&self, command: MakeMoveCommand);
+    fn xadd_provide_history(&self, command: ProvideHistoryCommand);
+    fn xadd_req_sync(&self, command: ReqSyncBackendCommand);
+    fn xadd_join_private_game(&self, command: JoinPrivateGameBackendCommand);
+    fn xadd_find_public_game(&self, command: FindPublicGameBackendCommand);
+    fn xadd_create_game(&self, command: CreateGameBackendCommand);
+    fn xadd_choose_color_pref(&self, command: ChooseColorPrefBackendCommand);
+    fn xadd_session_disconnected(&self, command: SessionDisconnected);
 }
 
-pub fn start(commands_out: Receiver<BackendCommands>, cmds: &dyn XAddCommands) {
+pub fn start(commands_out: Receiver<BC>, cmds: &dyn XAddCommands) {
     loop {
         select! {
             recv(commands_out) -> backend_command_msg => match backend_command_msg {
                 Err(e) => error!("backend command xadd {:?}",e),
-                Ok(command) => match command {
-                    BackendCommands::AttachBot(attach_bot) => {
-                        cmds.xadd_attach_bot(attach_bot)
+                Ok(command) => {
+                    match &command {
+                        BC::ClientHeartbeat(_) => (), // no one cares
+                        _ => info!("Send command: {:?}", &command)
                     }
-                    BackendCommands::MakeMove(c) => cmds.xadd_make_move(c),
-                    BackendCommands::ClientHeartbeat(_) => (),
-                    _ => error!("cannot match backend command to xadd"),
+                    match command {
+                        BC::AttachBot(attach_bot) => cmds.xadd_attach_bot(attach_bot),
+                        BC::MakeMove(c) => cmds.xadd_make_move(c),
+                        BC::ClientHeartbeat(_) => (),
+                        BC::ProvideHistory(ph) => cmds.xadd_provide_history(ph),
+                        BC::ReqSync(rs) => cmds.xadd_req_sync(rs),
+                        BC::JoinPrivateGame(jpg) => cmds.xadd_join_private_game(jpg),
+                        BC::FindPublicGame(fpg) => cmds.xadd_find_public_game(fpg),
+                        BC::CreateGame(cg) => cmds.xadd_create_game(cg),
+                        BC::ChooseColorPref(cp) => cmds.xadd_choose_color_pref(cp),
+                        BC::SessionDisconnected(sd) => cmds.xadd_session_disconnected(sd),
+                        _ => error!("cannot match backend command to xadd"),
+                    }
                 }
             }
         }
@@ -45,7 +68,7 @@ impl XAddCommands for RedisXAddCommands {
             Ok(bin) => {
                 let mut redis_cmd = redis::cmd("XADD");
                 redis_cmd
-                    .arg(ATTACH_BOT_TOPIC)
+                    .arg(topics::ATTACH_BOT_TOPIC)
                     .arg("MAXLEN")
                     .arg("~")
                     .arg("1000")
@@ -63,7 +86,7 @@ impl XAddCommands for RedisXAddCommands {
 
         let mut redis_cmd = redis::cmd("XADD");
         redis_cmd
-            .arg(MAKE_MOVE_TOPIC)
+            .arg(topics::MAKE_MOVE_TOPIC)
             .arg("MAXLEN")
             .arg("~")
             .arg("1000")
@@ -81,11 +104,84 @@ impl XAddCommands for RedisXAddCommands {
             error!("make move {:?}", e)
         }
     }
+
+    fn xadd_provide_history(&self, command: ProvideHistoryCommand) {
+        self.xadd_classic(
+            bincode::serialize(&command.into_shared()),
+            topics::PROVIDE_HISTORY_TOPIC,
+        )
+    }
+
+    fn xadd_join_private_game(&self, command: JoinPrivateGameBackendCommand) {
+        self.xadd_classic(
+            bincode::serialize(&command.into_shared()),
+            topics::JOIN_PRIVATE_GAME_TOPIC,
+        )
+    }
+
+    fn xadd_find_public_game(&self, command: FindPublicGameBackendCommand) {
+        self.xadd_classic(
+            bincode::serialize(&command.into_shared()),
+            topics::FIND_PUBLIC_GAME_TOPIC,
+        )
+    }
+
+    fn xadd_create_game(&self, command: CreateGameBackendCommand) {
+        self.xadd_classic(
+            bincode::serialize(&command.into_shared()),
+            topics::CREATE_GAME_TOPIC,
+        )
+    }
+
+    fn xadd_req_sync(&self, command: ReqSyncBackendCommand) {
+        self.xadd_classic(
+            bincode::serialize(&command.into_shared()),
+            topics::REQ_SYNC_TOPIC,
+        )
+    }
+
+    fn xadd_choose_color_pref(&self, command: ChooseColorPrefBackendCommand) {
+        self.xadd_classic(
+            bincode::serialize(&command.into_shared()),
+            topics::CHOOSE_COLOR_PREF_TOPIC,
+        )
+    }
+
+    fn xadd_session_disconnected(&self, command: SessionDisconnected) {
+        self.xadd_classic(
+            bincode::serialize(&command.into_shared()),
+            topics::SESSION_DISCONNECTED_TOPIC,
+        )
+    }
 }
 
 impl RedisXAddCommands {
     pub fn create(pool: Arc<RedisPool>) -> Self {
         RedisXAddCommands { pool }
+    }
+
+    fn xadd_classic(&self, bin: Result<Vec<u8>, Box<bincode::ErrorKind>>, topic: &str) {
+        match self.pool.get() {
+            Err(e) => error!("xadd {}: cannot get conn {:?}", topic, e),
+            Ok(mut conn) => {
+                if let Ok(b) = bin {
+                    let mut redis_cmd = redis::cmd("XADD");
+                    redis_cmd
+                        .arg(topic)
+                        .arg("MAXLEN")
+                        .arg("~")
+                        .arg("1000")
+                        .arg("*")
+                        .arg("data")
+                        .arg(b);
+                    if let Err(e) = redis_cmd.query::<String>(&mut *conn) {
+                        error!("xadd {}: redis execution err. {:?}", topic, e)
+                    }
+                } else {
+                    error!("xadd {}: serialization error", topic)
+                }
+            }
+        }
     }
 }
 
@@ -104,25 +200,65 @@ mod tests {
     enum TestResult {
         Bot(AttachBot),
         Move(MakeMoveCommand),
+        Hist(ProvideHistoryCommand),
+        RSyn(ReqSyncBackendCommand),
+        Join(JoinPrivateGameBackendCommand),
+        Find(FindPublicGameBackendCommand),
+        Create(CreateGameBackendCommand),
+        ChCol(ChooseColorPrefBackendCommand),
+        SessDisconn(SessionDisconnected),
+    }
+    impl FakeXAddCmd {
+        fn sssend(&self, tr: TestResult) {
+            self.st.send(tr).expect("send")
+        }
     }
     impl XAddCommands for FakeXAddCmd {
         fn xadd_attach_bot(&self, attach_bot: AttachBot) {
-            self.st.send(TestResult::Bot(attach_bot)).expect("send")
+            self.sssend(TestResult::Bot(attach_bot))
         }
         fn xadd_make_move(&self, command: MakeMoveCommand) {
-            self.st.send(TestResult::Move(command)).expect("send")
+            self.sssend(TestResult::Move(command))
+        }
+
+        fn xadd_provide_history(&self, command: ProvideHistoryCommand) {
+            self.sssend(TestResult::Hist(command))
+        }
+
+        fn xadd_join_private_game(&self, command: JoinPrivateGameBackendCommand) {
+            self.sssend(TestResult::Join(command))
+        }
+
+        fn xadd_find_public_game(&self, command: FindPublicGameBackendCommand) {
+            self.sssend(TestResult::Find(command))
+        }
+
+        fn xadd_create_game(&self, command: CreateGameBackendCommand) {
+            self.sssend(TestResult::Create(command))
+        }
+
+        fn xadd_req_sync(&self, command: ReqSyncBackendCommand) {
+            self.sssend(TestResult::RSyn(command))
+        }
+
+        fn xadd_choose_color_pref(&self, command: ChooseColorPrefBackendCommand) {
+            self.sssend(TestResult::ChCol(command))
+        }
+
+        fn xadd_session_disconnected(&self, command: SessionDisconnected) {
+            self.sssend(TestResult::SessDisconn(command))
         }
     }
 
     #[test]
     fn test_loop() {
         let (test_in, test_out): (Sender<TestResult>, Receiver<TestResult>) = unbounded();
-        let (cmds_in, cmds_out): (Sender<BackendCommands>, Receiver<BackendCommands>) = unbounded();
+        let (cmds_in, cmds_out): (Sender<BC>, Receiver<BC>) = unbounded();
 
         thread::spawn(move || start(cmds_out, &FakeXAddCmd { st: test_in }));
 
         cmds_in
-            .send(BackendCommands::AttachBot(AttachBot {
+            .send(BC::AttachBot(AttachBot {
                 game_id: micro_model_moves::GameId(Uuid::nil()),
                 board_size: Some(9),
                 player: micro_model_moves::Player::WHITE,
@@ -130,7 +266,7 @@ mod tests {
             .expect("send test");
 
         cmds_in
-            .send(BackendCommands::MakeMove(MakeMoveCommand {
+            .send(BC::MakeMove(MakeMoveCommand {
                 game_id: Uuid::nil(),
                 req_id: Uuid::nil(),
                 player: Player::BLACK,